Move VAB tests
